[IMP] sync with ONL
Multiple code formats and syncs with onlogic
