+++ ol_product_configurator/models/product_config.py
@@ -12,7 +12,7 @@ class ProductConfigSession(models.Model):
     # COLUMNS ##########
 
     company_id = fields.Many2one(
-        "res.company",
+        comodel_name="res.company",
         string="Company",
         default=lambda self: self.env.company,
     )
