+++ sale_substate/README.rst
@@ -7,7 +7,7 @@ Sale Sub State
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:30c35ddeeceae6b4a279d86652752cc3fbf10f0a9e73f8054ab7effd2ace4e99
+   !! source digest: sha256:27ee2f0bb5e10f298583dd5d40392621dd01590b77f02879021f9479fa696c62
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 .. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
@@ -72,8 +72,8 @@ Authors
 Contributors
 ------------
 
--  Mourad EL HADJ MIMOUNE <mourad.elhadj.mimoune@akretion.com>
--  Alexei Rivera <arivera@archeti.com> (migration to 15.0)
+- Mourad EL HADJ MIMOUNE <mourad.elhadj.mimoune@akretion.com>
+- Alexei Rivera <arivera@archeti.com> (migration to 15.0)
 
 Maintainers
 -----------
