+++ sale_substate/i18n/sale_substate.pot
@@ -4,7 +4,7 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 16.0\n"
+"Project-Id-Version: Odoo Server 17.0\n"
 "Report-Msgid-Bugs-To: \n"
 "Last-Translator: \n"
 "Language-Team: \n"
@@ -138,15 +138,6 @@ msgstr ""
 #. module: sale_substate
 #: model_terms:ir.ui.view,arch_db:sale_substate.view_sales_order_filter
 msgid "Substate"
-msgstr ""
-
-#. module: sale_substate
-#. odoo-python
-#: code:addons/sale_substate/models/sale_order.py:0
-#, python-format
-msgid ""
-"The substate %(name)s is not defined for the state %(state)s but for "
-"%(target_state)s "
 msgstr ""
 
 #. module: sale_substate
